fix(Phase 4): NumPy return type conversions for DSP utilities
Files fixed:
- auralis/dsp/utils/conversion.py (1 error)
  * Wrapped np.log10 return with float() in to_db()
  * Fixes: Returning Any from function declared to return "float"

- auralis/dsp/utils/audio_info.py (2 errors)
  * Wrapped audio.shape[0] with int() in size()
  * Wrapped audio.shape[1] with int() in channel_count()
  * Fixes: Returning Any from function declared to return "int"

Total errors fixed: 3 (from 313 to 310)
Mypy error count: 310 errors in 124 files (down from 313)
Phase 4 progress: 34 errors fixed total (9.9% improvement from baseline 344)

+++ auralis/dsp/utils/conversion.py
@@ -23,7 +23,7 @@ def to_db(linear_value: float) -> float:
     Returns:
         Value in decibels (dB)
     """
-    return 20 * np.log10(max(linear_value, 1e-10))
+    return float(20 * np.log10(max(linear_value, 1e-10)))
 
 
 def from_db(db_value: float) -> float:
